in SocialMediaController{ To start, i set it up as a @RestController, and initilized AccountService and MessageService.
i @autowired the constructor, allowing it to initilize correctly

from there i set up mapping for the 8 requested problems.

i postmapped /register, attempting to register the account with AccountService,
i would get back either a null, a Account with no Id or a full account

i postmapped /login, attempting to log in with AccountService
i would get back either a null, if there was a problem with acquisition
or an account, from there, i would check if the password was correct,
sending 200 and the account if it was, or 401 if it was wrong.

i postmapped /messages, and attempted to post a message,
if there was a problem i received a null, otherwise i received
the posted message and returned that with the 200 status.

i getmapped /messages, getting a list of all the messages,
i returned the list with status 200

i getmapped /messages/{messageId}, getting the specific
message, returning it with a 200 status.

i deletemapped /message/{messageId}, recieving a bool from MessageService
if it was successful, allowing me to change the body from 1 or nothing depending
on success.

i patchmapped /message/{messageId}, attempting to patch a message through
MessageService, if it was successful i got the new message, if not
i got null.

i getmapped /accounts/{accountId}/messages attempting to get
a list of all of an accounts messages returning the list with a
200 status.
}

in MessageService{
To start, i set it up as a @Service ,and initilized MessageRepository
and AccountRepository then @autowired the constructor.

I set up a @Transactional for postMessage, and verified data, during which
i used the only use of AccountRepository in MessageService to validate an account
if the data was valid, i returned the saved message.

for getMessages i returned MessageRepository.getAllMessages

for getMessageById i returned MessageRepository.getMessageById

for deleteMessageById i setup a @Transactional, from there i checked if
the message was there to be deleted, if it was i deleteed it.

for getMessageByAccount i returned MessageRepository.findByAccountId

for patchMessage i setup a @Transactional, from there i validated the data, and if the
data was valid, i updated the data, returning the new message if it was valid.
}

in MessageRepository{
to start, i set it up as a @Repository

i set up a @Query for findByAccountId, making a select statment to
search for all messages posted by an account

i setup a @Query for getAllMessages, tha made a list of all messaged in the database

i setup a @Query for getMessageById, that got the one message from
the list that shared the message Id.

i setup a @Query with a @Modifying for deleteMessageById that
deleted the message that had a specific messageId

i setup a @Query with a @Modifying for patchMessageById that
updated a message with a specific messageId with a new messageText
}

in AccountService{
to start, i set it up as a @Service, and initilized AccountRepository

from there i @autowired the constructor

for registerAccount, i checked for valitity, returning null if there was a
problem with the username or password, if the username was taken, i returned the
account without the id it gets from being inserted, if the data was all
correct, and avalible, i returned AccountRepository.save

for loginAccount, i returned AccountRepository.findByUsername
}
in AccountRepository{
to start, i set it up as a @Repository

i setup findByUsername to look through the accounts for a specific
user, returning them if possible.

i setup findAccountByMessageId for MessageService to Query for a specific
account.

}